System environment variables no longer override local environment variables.
This patch addresses:
https://github.com/cbucher/console/issues/345

The call to CreateEnvironmentBlock seems to be the issue.  Even though TRUE was passed in asking to inherit the environment from the parent process, that doesn't seem to mean what I think it means.  If the parent process and system environment have different values for an environment variable XYZ, then the system one is returned in this logic.  However, the GetEnvironmentStrings method seems to do the right thing.